refactor onboarding bottom component for improved layout and styling
